+++ database/database.js
@@ -1,17 +1,20 @@
-const read_graphy = require('graphy').content.nt.read;
+
 const dataset_tree = require('graphy').util.dataset.tree
 const factory = require('@graphy/core.data.factory');
 const MongoClient = require('mongodb').MongoClient;
 const jsonld = require('jsonld');
+const databaseUtilities = require('./databaseUtilities')
+const mongodbUtilities = require('./mongodb/Utilities')
 
 // IN URI OR LITERAL -> OUT -> Literal or URI or Quad or Boolean
 class Database {
     constructor(schemaMapping) {
+        databaseUtilities.createReverseContext(schemaMapping)
         this.schemaMapping = schemaMapping;
+
         this.database = dataset_tree();
         this.stampleDataType = "http://staple-api.org/datamodel/type";
         this.pages = [];
-        // this.insertFakeDataToDB = false;
     }
 
     // ---
@@ -24,18 +27,12 @@ class Database {
         gra = factory.namedNode(gra);
 
         let quad = factory.quad(sub, pred, obj, gra);
-
         this.database.add(quad);
-        // if(this.insertFakeDataToDB === false){
-        //     this.insertFakeDataToDB = true;
-        //     this.mongodbAdd();
-        // }
         return true;
     }
 
     // ---  
     delete(sub, pred, obj, gra = null) {
-
         sub = factory.namedNode(sub);
         pred = factory.namedNode(pred);
         if (typeof (obj) !== "object" && obj !== undefined) {
@@ -85,21 +82,8 @@ class Database {
         return removed;
     }
 
-    findContextName(uri) {
-        for (let key in this.schemaMapping["@context"]) {
-            if (this.schemaMapping["@context"][key] === uri) {
-                return key;
-            }
-        }
-        return uri;
-    }
-
     // Array of uri
     async getObjectsValueArray(sub, pred, expectLiterals = false) {
-        // console.log("getObjectsValueArray looking for :")
-        // console.log(sub)
-        // console.log(pred)
-
         sub = factory.namedNode(sub);
         pred = factory.namedNode(pred);
 
@@ -116,15 +100,11 @@ class Database {
             }
             x = itr.next();
         }
-        // console.log("Found:")
-        // console.log(data)
-        // console.log(" ")
         return data;
     };
 
     // Array of uri
     isTripleInDB(sub, pred, obj, gra = null) {
-
         sub = factory.namedNode(sub);
         pred = factory.namedNode(pred);
         if (typeof (obj) !== "object" && obj !== undefined) {
@@ -134,13 +114,11 @@ class Database {
 
 
         let quad = factory.quad(sub, pred, obj, gra);
-
         return this.database.has(quad)
     };
 
     // Array of Quads
     getTriplesBySubject(sub) {
-
         sub = factory.namedNode(sub);
 
         const temp = this.database.match(sub, null, null);
@@ -157,6 +135,7 @@ class Database {
     // Array of Quads
     getTriplesByObjectUri(uri) {
         uri = factory.namedNode(uri);
+
         const temp = this.database.match(null, null, uri);
         let data = [];
         var itr = temp.quads();
@@ -170,20 +149,17 @@ class Database {
 
     // returns single object value - uri or data
     getSingleStringValue(sub, pred) {
-
         sub = factory.namedNode(sub);
         pred = factory.namedNode(pred);
 
         const temp = this.database.match(sub, pred, null);
         var itr = temp.quads();
         var x = itr.next();
-
         return x.value.object.value;
     };
 
     // returns single object value - data
     getSingleLiteral(sub, pred) {
-
         sub = factory.namedNode(sub);
         pred = factory.namedNode(pred);
 
@@ -198,16 +174,10 @@ class Database {
         return x.value.object;
     };
 
-    // returns array of uri
+    // returns array of uri - Core Query
     async getSubjectsByType(type, predicate, inferred = false, page = undefined, query = undefined) {
+        await this.loadCoreQueryDataFromDB(type, page, query, inferred);
 
-
-        console.log("Before")
-        console.log(this.database.size)
-        console.log(`Asking for type ${type} page ${page}`)
-        await this.loadCoreQueryDataFromDB(type, page, query, inferred)
-        console.log("After")
-        console.log(this.database.size)
         type = factory.namedNode(type);
 
         if (inferred) {
@@ -246,71 +216,6 @@ class Database {
         this.database.clear();
     }
 
-    insertRDFPromise(tree, ID, rdf) {
-        return new Promise((resolve, reject) => {
-            let data = (y_quad) => {
-                if(y_quad.subject.value ===  ID){
-                    console.log(y_quad.subject.value + " === " + ID)
-                    console.log(`predicate ${y_quad.predicate.value}`)
-                    console.log(`object ${y_quad.object.value}`)
-
-                }
-                if (y_quad.subject.value ===  ID) {
-                    y_quad.graph = factory.namedNode(null);
-
-                    // add inverses 
-                    let inverse = this.schemaMapping['@graph'].filter(x => x["@id"] === y_quad.predicate.value)
-                        // console.log("inverse") 
-                        // console.log(y_quad)
-                        // console.log(inverse)
-                        // console.log("\n")
-                    inverse = inverse[0]
-                    if (inverse !== undefined) {
-                        inverse['http://schema.org/inverseOf'].forEach(inversePredicate => {
-                            let quad = factory.quad(y_quad.object, factory.namedNode(inversePredicate), y_quad.subject, y_quad.graph);
-                            tree.add(quad);
-                        })
-                    }
-
-                    tree.add(y_quad);
-                }
-            }
-
-            let eof = (h_prefixes) => {
-                resolve('done')
-            }
-
-            read_graphy(rdf, { data, eof, })
-        });
-    }
-
-    async insertRDF(rdf, ID) {
-        await this.insertRDFPromise(this.database, ID, rdf);
-        this.updateInference();
-    }
-
-    removeRDFPromise(tree, ID, rdf) {
-        return new Promise((resolve, reject) => {
-            let data = (y_quad) => {
-                if (y_quad.subject.value === ID) {
-                    y_quad.graph = factory.namedNode(null);
-                    tree.delete(y_quad);
-                }
-            }
-
-            let eof = (h_prefixes) => {
-                resolve('done')
-            }
-
-            read_graphy(rdf, { data, eof, })
-        });
-    }
-
-    async removeRDF(rdf, ID) {
-        await this.removeRDFPromise(this.database, ID, rdf);
-        this.updateInference();
-    }
-
     updateInference() {
         // remove all staple : datatype but not Thing 
         let temp = this.database.match(null, null, null);
@@ -327,251 +232,45 @@ class Database {
         temp = this.database.match(null, null, null);
         itr = temp.quads();
         itrData = itr.next();
-        let addedQuads = []
 
         while (!itrData.done) {
-
             if (itrData.value.predicate.value === "http://www.w3.org/1999/02/22-rdf-syntax-ns#type") {
                 let data = this.schemaMapping["@graph"].filter((x) => { return x['@id'] === itrData.value.object.value })
                 for (let key in data) {
                     let uris = data[key]["http://www.w3.org/2000/01/rdf-schema#subClassOf"];
                     for (let x in uris) {
                         this.create(itrData.value.subject.value, this.stampleDataType, uris[x]['@id'])
-                        addedQuads.push(`${itrData.value.subject.value}, ${this.stampleDataType}, ${uris[x]['@id']}`)
                     }
-
                 }
             }
             itrData = itr.next();
         }
-
-    }
-
-    // Based on reverse property
-    async loadChildObjectsFromDB(sub, pred, type) {
-        let url = 'mongodb://127.0.0.1:27017';
-        const client = await MongoClient.connect(url, { useNewUrlParser: true })
-            .catch(err => { console.log(err); });
-
-        if (!client) {
-            return;
-        }
-
-        try {
-            console.log("START CONNECTION")
-            const db = client.db("staple");
-            let collection = db.collection('quadsTEST');
-            let query = { _type: 'Thing' }
-
-            if (type != undefined) {
-                query = { _type: type }
-            }
-
-            for (let key in this.schemaMapping['@context']) {
-                if (this.schemaMapping['@context'][key] === pred) {
-                    query['_reverse'] = {};
-                    query['_reverse'][key] = [{ '_id': sub }]
-                    break;
-                }
-            }
-
-
-            // console.log(`finall query['_reverse']: `)
-            // console.log(query['_reverse'])
-
-            let result = await collection.find(query).toArray();
-
-            result = result.map(x => {
-                x['@context'] = this.schemaMapping['@context'];
-                return x;
-            })
-
-            const rdf = await jsonld.toRDF(result, { format: 'application/n-quads' });
-            for (let obj in result) {
-                await this.insertRDF(rdf, result[obj]['_id']);
-            }
-        } catch (err) {
-            console.log(err);
-        } finally {
-            client.close();
-        }
-    }
-
-    async loadCoreQueryDataFromDB(type, page = 1, query = undefined, inferred = false) {
-        let url = 'mongodb://127.0.0.1:27017';
-        const client = await MongoClient.connect(url, { useNewUrlParser: true })
-            .catch(err => { console.log(err); });
-
-        if (!client) {
-            return;
-        }
-
-        try {
-            const db = client.db("staple");
-            let collection = db.collection('quadsTEST');
-            let _type = undefined;
-
-            if (query === undefined) {
-                for (let key in this.schemaMapping['@context']) {
-                    if (this.schemaMapping['@context'][key] === type) {
-                        _type = key;
-                        break;
-                    }
-                }
-            }
-
-            if (_type !== undefined) {
-                if (inferred) {
-                    query = { _inferred: _type }
-                }
-                else {
-                    query = { _type: _type }
-                }
-            }
-
-            let result;
-            if (page === undefined) {
-                result = await collection.find(query).toArray();
-            }
-            else {
-                result = await collection.find(query).skip(page * 10 - 10).limit(10).toArray();
-            }
-
-            // save page conetnt
-            this.pages[page] = result.map(x => x['_id'])
-
-            result = result.map(x => {
-                x['@context'] = this.schemaMapping['@context'];
-                return x;
-            })
-
-            const rdf = await jsonld.toRDF(result, { format: 'application/n-quads' });
-            for (let obj in result) {
-                console.log(result[obj]['_id'])
-                await this.insertRDF(rdf, result[obj]['_id']);
-            }
-        } catch (err) {
-            console.log(err);
-        } finally {
-            client.close();
-        }
-
-    }
-
-    mongodbAddOrUpdate(flatJson) {
-        MongoClient.connect('mongodb://127.0.0.1:27017', async function (err, db) {
-            if (err) {
-                throw err;
-            } else {
-                var dbo = db.db("staple");
-                let collection = dbo.collection('quadsTEST');
-
-                let result = await collection.find({ "_id": flatJson['_id'] }).toArray();
-
-                if (result[0] !== undefined) {
-                    collection.update({ _id: flatJson['_id'] }, flatJson);
-                }
-                else {
-                    dbo.collection("quadsTEST").insertOne(flatJson, function (err, res) {
-                        if (err) throw err;
-                        console.log("quad inserted");
-                    });
-                }
-                db.close();
-            }
-        })
     }
 
     async getFlatJson() {
-        // find all objects
-        let ids = await this.getSubjectsByType("http://schema.org/Thing", this.stampleDataType);
+        return await databaseUtilities.getFlatJson(this);
+    }
 
-        // CHECK IF OBJECT ALREADY IN DATABASE IF SO, THEN UPDATE
+    async loadChildObjectsFromDB(sub, pred, type) {
+        await mongodbUtilities.loadChildObjectsFromDB(this, sub, pred, type)
+    }
 
-        for (let i in ids) {
-            let id = ids[i]
-            let allRelatedQuads = []
+    async loadCoreQueryDataFromDB(type, page = 1, query = undefined, inferred = false) {
+        await mongodbUtilities.loadCoreQueryDataFromDB(this, type, page, query, inferred);
+    }
 
-            var temp = this.database.match(factory.namedNode(id), null, null);
-            var itr = temp.quads();
-            var x = itr.next();
-            while (!x.done) {
-                allRelatedQuads.push(x.value);
-                x = itr.next();
-            }
+    mongodbAddOrUpdate(flatJson) {
+        mongodbUtilities.mongodbAddOrUpdate(flatJson);
+    }
 
-            temp = this.database.match(null, null, factory.namedNode(id));
-            itr = temp.quads();
-            x = itr.next();
-            while (!x.done) {
-                allRelatedQuads.push(x.value);
-                x = itr.next();
-            }
+    async insertRDF(rdf, ID) {
+        await databaseUtilities.insertRDFPromise(this.database, ID, rdf, this.schemaMapping);
+        this.updateInference();
+    }
 
-            // console.log(allRelatedQuads)
-            //create json
-            let newJson = {
-                "_id": id,
-                "_type": undefined,
-                "_inferred": [],
-                "_reverse": {},
-            }
-
-            for (let quad in allRelatedQuads) {
-                quad = allRelatedQuads[quad]
-                // console.log(quad)
-                if (quad.subject.value === id) {
-                    if (quad.predicate.value === "http://www.w3.org/1999/02/22-rdf-syntax-ns#type") {
-                        let contextKey = this.findContextName(quad.object.value)
-                        newJson["_type"] = contextKey;
-                    }
-                    else if (quad.object.datatype !== undefined) { //Literal
-                        //find key 
-                        let fieldKey = quad.predicate.value;
-                        if (quad.predicate.value === this.schemaMapping["@context"][quad.predicate.value.split("http://schema.org/")[1]]) {
-                            fieldKey = quad.predicate.value.split("http://schema.org/")[1];
-                        }
-                        else {
-                            fieldKey = this.findContextName(quad.predicate.value);
-                        }
-                        if (newJson[fieldKey] === undefined) {
-                            newJson[fieldKey] = []
-                        }
-
-                        newJson[fieldKey].push({
-                            _value: quad.object.value,
-                            _type: this.findContextName(quad.object.datatype.value),
-                        });
-                    }
-                    else if (quad.predicate.value === "http://staple-api.org/datamodel/type") { // _inferred
-                        let contextKey = this.findContextName(quad.object.value);
-                        newJson['_inferred'].push(contextKey)
-                    }
-                    else { // object
-                        let contextKey = this.findContextName(quad.predicate.value);
-
-                        if (newJson[contextKey] === undefined) {
-                            newJson[contextKey] = []
-                        }
-
-                        newJson[contextKey].push({ _id: quad.object.value });
-                    }
-                }
-
-                if (quad.object.value === id) { // _reverse
-                    let contextKey = this.findContextName(quad.predicate.value);
-
-                    if (newJson['_reverse'][contextKey] === undefined) {
-                        newJson['_reverse'][contextKey] = []
-                    }
-
-                    newJson['_reverse'][contextKey].push({ _id: quad.subject.value })
-                }
-            }
-            //add to mongo db
-            this.mongodbAddOrUpdate(newJson)
-        }
-        return "DONE"
+    async removeRDF(rdf, ID) {
+        await databaseUtilities.removeRDFPromise(this.database, ID, rdf);
+        this.updateInference();
     }
 }
